Add support for reproducible builds to setuptools/wheel
**What is the current situation/context?**

Currently the combination of setuptools/wheel does not support
reproducible builds completely.

**What ought to be fixed, made, or implemented?**

Complete support for reproducible build for both sdist and wheel using
setuptools as a build backend.

**What problems would this solve, and what new capabilities would it cause?**

This would help to improve security in the Python packaging ecosystem,
because developers would be able to independently verify packages.o

**What kinds of work are necessary to make this happen?**

Implementation efforts are required for:

- Support `SOURCE_DATA_EPOCH` environment variable for both sdists and wheels.
- Make both sdist and wheels independent of umask.
- Ensure that C/C++-extensions compiled with setuptools are reproducible.

Documentation efforts are required to instruct developers how to verify
distribution artifacts.